Half way through term unification algorithm.
I've written a partial algorithm, but it won't work at all. I need to work out
a general way of unifying graphs, and in order to do that I need to be able to
identify the set of entry points to the term graph in a repeatable way,
otherwise I'll have to store every possible entry point.