Added memcpy instruction, modified alloca and added byte type.
byte type is intended to be similar to Cs void*; it can be used to copy data
without any knowledge of its contents. I might remove this later since
integers of various sizes cover this case and I might need the multiple sizes
to cover different alignments.

I've added an alignment parameter and a number of elements parameter to alloca
to support aggregate lowering.