Android TVM testing appears to be written correctly, but fails.
The test program ends up in an infinite loop, which appears to be
cause by LLVM failing to switch the permissions for the generated
code from RW to RX permissions. For some reason, this gives an
infinite loop rather than a segfault (Dalvik signal handlers?).

I've tested this on a real Galaxy Nexus, which gives the same
behaviour as the emulator.