Binary size reduction optimizations.
I think this is about as small as I can make the various binaries as-is.
Merging shared object will reduce the total size a bit further, but really
I need to write an alternate JIT backend to produce a much smaller footprint.

For some reason Win32 tests are failing (at least under Wine). I don't know
at the moment whether this is because I'm mis-generating the LLVM IR;
it'll be much easier to debug from real Windows.